More work on the website. You can now add and delete people on the person table.

+++ src/main/webapp/assets/javascript/rest.js
@@ -5,26 +5,17 @@
  */
 
 
-//JQuery
-$(document).ready(function ()
+
+window.addEventListener("load", function ()
 {
+
     PersonsRefresh();
+    CompaniesRefresh();
 
-    $("#PersonsRefresh").click(function ()
-    {
-        PersonsRefresh();
-    });
-    
-    CompanyRefresh();
-
-    $("#CompaniesRefresh").click(function ()
-    {
-        CompaniesRefresh();
-    });
 
 });
 
-
+//JQuery / Javascript
 
 function PersonsRefresh()
 {
@@ -32,26 +23,31 @@ function PersonsRefresh()
     {
         $('#TablePersons tr').remove();
         var rows = "";
-        rows += '<tr><th>' + "ID" + '</th><th>' + "First name" + '</th><th>' + "Last name" + '</th><th>' + "Street" + '</th><th>' + "Zipcode" + '</th><th>' + "City" + '</th><th>' + "DELETE ROW" + '</th></tr>';
+
+        rows += '<tr><th>' + "SELECT" + '</th><th>' + "ID" + '</th><th>' + "First name" + '</th><th>' + "Last name" + '</th><th>' + "Street" + '</th><th>' + "HouseNr" + '</th><th>' + "Zipcode" + '</th><th>' + "City" + '</th></tr>';
         $.each(persons, function (i, person)
         {
-            rows += '<tr><td>' + person.id + '</td><td>' + person.firstname + '</td><td>' + person.lastname + '</td><td>' + person.address + '</td><td>' + person.zip + '</td><td>' + person.city + '</td><td><input type="button" onclick="PersonDelete(' + person.id + ')" value="Delete" /></td></tr>';
+            rows += '<tr><td><input type="checkbox" class="checkmarkPerson" name="rowCheck" value="' + person.id + '" > ' + '</td><td>' + person.id + '</td><td>' + person.firstname + '</td><td>' + person.lastname + '</td><td>' + person.address.street + '</td><td>' + person.address.additionalinfo + '</td><td>' + person.zip + '</td><td>' + person.city + '</td></tr>';
         });
-
         $('#TablePersons').append(rows);
+
     });
 }
 
 function PersonDelete(id)
 {
-    $.ajax({
-        url: "api/person/delete/" + id,
-        type: "DELETE",
-        data: "",
-        dataType: "json",
-        contentType: "application/json; charset=utf-8",
-        success: function () {}
-    });
+    var r = confirm("Are you sure you want to delete this row?");
+    if (r === true)
+    {
+        $.ajax({
+            url: "api/person/delete/" + id,
+            type: "DELETE",
+            data: "",
+            dataType: "json",
+            contentType: "application/json; charset=utf-8",
+            success: function () {}
+        });
+    }
 }
 
 function CompaniesRefresh()
@@ -60,10 +56,10 @@ function CompaniesRefresh()
     {
         $('#TableCompanies tr').remove();
         var rows = "";
-        rows += '<tr><td>' + "ID" + '</td><td>' + "First name" + '</td><td>' + "Last name" + '</td><td>' + "Street" + '</td><td>' + "Zipcode" + '</td><td>' + "City" + '</td></tr>';
+        rows += '<tr><th>' + "SELECT" + '</th><td>' + "ID" + '</td><td>' + "First name" + '</td><td>' + "Last name" + '</td><td>' + "Street" + '</td><td>' + "Zipcode" + '</td><td>' + "City" + '</td></tr>';
         $.each(companies, function (i, company)
         {
-           rows += '<tr><td>' + company.id + '</td><td>' + company.firstname + '</td><td>' + company.lastname + '</td><td>' + company.address + '</td><td>' + company.zip + '</td><td>' + company.city + '</td><td><input type="button" onclick="PersonDelete(' + company.id + ')" value="Delete" /></td></tr>';
+            rows += '<tr><td><input type="checkbox" class="checkmarkCompany" name="rowCheck" value="person.id" /> ' + '</td><td>' + company.id + '</td><td>' + company.firstname + '</td><td>' + company.lastname + '</td><td>' + company.address + '</td><td>' + company.zip + '</td><td>' + company.city + '</td></tr>';
         });
 
         $('#TableCompanies').append(rows);
@@ -78,6 +74,159 @@ function CompanyDelete(id)
         data: "",
         dataType: "json",
         contentType: "application/json; charset=utf-8",
-        success: function () {}
+        success: function ()
+        {
+        }
     });
+    PersonsRefresh();
 }
+
+
+$("#deletePersonBtn").click(function ()
+{
+
+    $('[type="checkbox"].checkmarkPerson').each(function () {
+        var row = (this.checked ? $(this).val() : "");
+
+
+        //now test by coloring the row:
+        if (this.checked === true) {
+            console.log(row.valueOf());
+            PersonDelete(row.valueOf());
+        }
+    });
+
+
+});
+
+
+
+$("#addPersonBtn").click(function ()
+{
+
+    var rows = '<tr class="tmpPersonRow"><td>' + '</td><td>' + '</td><td>' + '<input type="text" name="firstname" required />' + '</td><td>' + '<input type="text" name="lastname" required />' + '</td><td>' + '<input type="text" name="street" required />' + '</td><td>' + '<input type="text" name="additionalinfo" required />' + '</td><td>' + '<input type="number" name="zipcode" min="500" required />' + '</td><td>' + '<input type="text" name="city" required />' + '</td></tr>';
+    $('#TablePersons').append(rows);
+
+    $("#pSubmit").remove();
+    $("#pCancel").remove();
+    $('#personBtns').append('<button type="submit" id="pSubmit" value="" onclick="TablePersonAdd();"><img src="assets/images/applyicon.png" class="button-table1" alt="applyicon.png"/></button>');
+    $('#personBtns').append(" ");
+    $('#personBtns').append('<button type="submit" id="pCancel" value="" onclick="TablePersonRemove();"><img src="assets/images/cancelicon.png" class="button-table1" alt="cancelicon.png"/></button>');
+
+});
+
+$("#addCompanyBtn").click(function ()
+{
+    var rows = '<tr class="tmpCompanyRow"><td>' + '</td><td>' + '</td><td>' + '<input type="text" name="company" required />' + '</td><td>' + '<input type="text" name="description" required />' + '</td><td>' + '<input type="number" name="marketvalue" required />' + '</td><td>' + '<input type="number" name="employees" min="1" required />' + '</td><td>' + '<input type="text" name="city" required />' + '</td></tr>';
+    $('#TableCompanies').append(rows);
+    
+    $("#cSubmit").remove();
+    $("#cCancel").remove();
+    $('#companyBtns').append('<button type="submit" id="cSubmit" value="" onclick="TableCompanyAdd();"><img src="assets/images/applyicon.png" class="button-table1" alt="applyicon.png"/></button>');
+    $('#companyBtns').append(" ");
+    $('#companyBtns').append('<button type="submit" id="cCancel" value="" onclick="TableCompanyRemove();"><img src="assets/images/cancelicon.png" class="button-table1" alt="cancelicon.png"/></button>');
+
+});
+
+function TablePersonAdd()
+{
+    $('.tmpPersonRow').each(function (i, obj) {
+
+        $firstname = $(this).find('input[name*="firstname"]'),
+                $lastname = $(this).find('input[name*="lastname"]'),
+                $street = $(this).find('input[name*="street"]'),
+                $houseNr = $(this).find('input[name*="additionalinfo"]'),
+                $zipcode = $(this).find('input[name*="zipcode"]'),
+                $city = $(this).find('input[name*="city"]');
+
+        var person = '{ firstname: ' + $firstname.val() + ', lastname: ' + $lastname.val()
+                + ',email: ' + 'an@email.dk'
+                + ',address:' + '{'
+                + 'additionalinfo: ' + $houseNr.val()
+                + ',street: ' + $street.val()
+                + '}'
+                + ', city: ' + $city.val()
+                + ', zip: ' + $zipcode.val()
+                + ',phonenumbers:' + '[]'
+                + ',phonenumbers.number: ' + "12345678"
+                + ',phonenumbers.description: ' + "asasdasd"
+                + ',hobbies:' + '[]'
+                + ',hobbies.description: ' + "thisHobby"
+                + ',hobbies.name: ' + "aHobby" + '}';
+
+        $.ajax({
+            url: "api/person/complete/add",
+            type: "POST",
+            data: person,
+            dataType: "json",
+            contentType: "application/json; charset=utf-8",
+            success: function () {}
+        });
+
+    });
+
+    PersonsRefresh();
+}
+
+function TablePersonRemove()
+{
+    $(".tmpPersonRow").remove();
+    $("#pSubmit").remove();
+    $("#pCancel").remove();
+}
+
+function TableCompanyRemove()
+{
+    $(".tmpCompanyRow").remove();
+    $("#cSubmit").remove();
+    $("#cCancel").remove();  
+}
+
+
+function TableCompanyAdd()
+{
+    $('.tmpCompanyRow').each(function (i, obj)
+    {
+
+        $firstname = $(this).find('input[name*="firstname"]'),
+                $lastname = $(this).find('input[name*="lastname"]'),
+                $street = $(this).find('input[name*="street"]'),
+                $houseNr = $(this).find('input[name*="additionalinfo"]'),
+                $zipcode = $(this).find('input[name*="zipcode"]'),
+                $city = $(this).find('input[name*="city"]');
+
+        var person = '{ firstname: ' + $firstname.val() + ', lastname: ' + $lastname.val()
+                + ',email: ' + 'an@email.dk'
+                + ',address:' + '{'
+                + 'additionalinfo: ' + $houseNr.val()
+                + ',street: ' + $street.val()
+                + '}'
+                + ', city: ' + $city.val()
+                + ', zip: ' + $zipcode.val()
+                + ',phonenumbers:' + '[]'
+                + ',phonenumbers.number: ' + "12345678"
+                + ',phonenumbers.description: ' + "asasdasd"
+                + ',hobbies:' + '[]'
+                + ',hobbies.description: ' + "thisHobby"
+                + ',hobbies.name: ' + "aHobby" + '}';
+
+        $.ajax({
+            url: "api/person/complete/add",
+            type: "POST",
+            data: person,
+            dataType: "json",
+            contentType: "application/json; charset=utf-8",
+            success: function () {}
+        });
+
+    });
+
+    CompaniesRefresh();
+}
+
+function TableCompanyRemove()
+{
+    $(".tmpCompanyRow").remove();
+}
+
+
